Combine snake_oil and snake_oil_field
