+++ test-data/ert/snake_oil/fields/permx0.grdecl
