+++ test-data/ert/snake_oil/fields/permx1.grdecl
