+++ test-data/ert/snake_oil/fields/permx2.grdecl
