+++ test-data/ert/snake_oil/fields/permx3.grdecl
