+++ test-data/ert/snake_oil/fields/permx4.grdecl
