+++ test-data/ert/snake_oil/fields/permx5.grdecl
