+++ test-data/ert/snake_oil/fields/permx6.grdecl
