+++ test-data/ert/snake_oil/fields/permx7.grdecl
