+++ test-data/ert/snake_oil/fields/permx8.grdecl
