+++ test-data/ert/snake_oil/fields/permx9.grdecl
