+++ test-data/ert/snake_oil/fields/poro2.grdecl
