+++ test-data/ert/snake_oil/fields/poro4.grdecl
