+++ test-data/ert/snake_oil/fields/poro6.grdecl
