+++ test-data/ert/snake_oil/forward_models/snake_oil_simulator.py
@@ -7,16 +7,27 @@ import resfo
 from oil_reservoir_synthesizer import OilSimulator
 
 
-def globalIndex(i, j, k, nx=10, ny=10, nz=10):
+def globalIndex(i, j, k, nx=10, ny=10):
     return i + nx * (j - 1) + nx * ny * (k - 1)
 
 
-def readParameters(filename):
+def read_seed(filename):
     params = {}
     with Path(filename).open(encoding="utf-8") as f:
         for line in f:
-            key, value = line.split(":", 1)
+            key, value = line.split(":")
             params[key] = value.strip()
+
+    return params
+
+
+def read_parameters(filename):
+    params = {}
+    with Path(filename).open(encoding="utf-8") as f:
+        for line in f:
+            key, value = line.split(" ")
+            _, name = key.split(":")
+            params[name] = value.strip()
 
     return params
 
@@ -172,8 +183,8 @@ def roundedInt(value):
 
 
 if __name__ == "__main__":
-    seed = int(readParameters("seed.txt")["SEED"])
-    parameters = readParameters("snake_oil_params.txt")
+    seed = int(read_seed("seed.txt")["SEED"])
+    parameters = read_parameters("parameters.txt")
 
     op1_divergence_scale = float(parameters["OP1_DIVERGENCE_SCALE"])
     op2_divergence_scale = float(parameters["OP2_DIVERGENCE_SCALE"])
