+++ test-data/ert/snake_oil/grid/CASE.EGRID
