+++ test-data/ert/snake_oil/observations/observations_loc.txt
