+++ test-data/ert/snake_oil/parameters/snake_oil_parameters_OP2.txt
