+++ test-data/ert/snake_oil/snake_oil_surface.ert
