--- test-data/ert/snake_oil_field/SEED
@@ -1,1 +1,0 @@
-67 -110 42 -42 106 34 96 0 18 42 76 -69 44 -40 -78 -61
