--- test-data/ert/snake_oil_field/forward_models/SNAKE_OIL_DIFF
@@ -1,4 +1,0 @@
-STDOUT snake_oil_diff.stdout
-STDERR snake_oil_diff.stderr
-
-EXECUTABLE snake_oil_diff.py
