--- test-data/ert/snake_oil_field/forward_models/SNAKE_OIL_NPV
@@ -1,4 +1,0 @@
-STDOUT snake_oil_npv.stdout
-STDERR snake_oil_npv.stderr
-
-EXECUTABLE snake_oil_npv.py
