--- test-data/ert/snake_oil_field/forward_models/SNAKE_OIL_SIMULATOR
@@ -1,4 +1,0 @@
-STDOUT snake_oil.stdout
-STDERR snake_oil.stderr
-
-EXECUTABLE snake_oil_simulator.py
