--- test-data/ert/snake_oil_field/observations/observations.txt
@@ -1,54 +1,0 @@
-SUMMARY_OBSERVATION WOPR_OP1_9
-{
-    VALUE   = 0.1;
-    ERROR   = 0.05;
-    DATE    = 2010-03-31;  -- (RESTART = 9)
-    KEY     = WOPR:OP1;
-};
-
-SUMMARY_OBSERVATION WOPR_OP1_36
-{
-    VALUE   = 0.7;
-    ERROR   = 0.07;
-    DATE    = 2010-12-26;  -- (RESTART = 36)
-    KEY     = WOPR:OP1;
-};
-
-SUMMARY_OBSERVATION WOPR_OP1_72
-{
-    VALUE   = 0.5;
-    ERROR   = 0.05;
-    DATE    = 2011-12-21;  -- (RESTART = 72)
-    KEY     = WOPR:OP1;
-};
-
-SUMMARY_OBSERVATION WOPR_OP1_108
-{
-    VALUE   = 0.3;
-    ERROR   = 0.075;
-    DATE    = 2012-12-15; -- (RESTART = 108)
-    KEY     = WOPR:OP1;
-};
-
-SUMMARY_OBSERVATION WOPR_OP1_144
-{
-    VALUE   = 0.2;
-    ERROR   = 0.035;
-    DATE    = 2013-12-10;  -- (RESTART = 144)
-    KEY     = WOPR:OP1;
-};
-
-SUMMARY_OBSERVATION WOPR_OP1_190
-{
-    VALUE   = 0.015;
-    ERROR   = 0.01;
-    DATE    = 2015-03-15;  -- (RESTART = 190)
-    KEY     = WOPR:OP1;
-};
-
-GENERAL_OBSERVATION WPR_DIFF_1 {
-   DATA       = SNAKE_OIL_WPR_DIFF;
-   INDEX_LIST = 400,800,1200,1800;
-   RESTART    = 199;
-   OBS_FILE   = wpr_diff_obs.txt;
-};
