--- test-data/ert/snake_oil_field/observations/wpr_diff_obs.txt
@@ -1,4 +1,0 @@
-0.0 0.1
-0.1 0.2
-0.2 0.15
-0.0 0.05
