--- test-data/ert/snake_oil_field/templates/seed_template.txt
@@ -1,1 +1,0 @@
-SEED:<IENS>
