--- test-data/ert/snake_oil_field/templates/snake_oil_template.txt
@@ -1,10 +1,0 @@
-OP1_PERSISTENCE:<OP1_PERSISTENCE>
-OP1_OCTAVES:<OP1_OCTAVES>
-OP1_DIVERGENCE_SCALE:<OP1_DIVERGENCE_SCALE>
-OP1_OFFSET:<OP1_OFFSET>
-OP2_PERSISTENCE:<OP2_PERSISTENCE>
-OP2_OCTAVES:<OP2_OCTAVES>
-OP2_DIVERGENCE_SCALE:<OP2_DIVERGENCE_SCALE>
-OP2_OFFSET:<OP2_OFFSET>
-BPR_555_PERSISTENCE:<BPR_555_PERSISTENCE>
-BPR_138_PERSISTENCE:<BPR_138_PERSISTENCE>
